auth-service/requirements.txt
flask==2.0.1
flask-sqlalchemy==2.5.1
pymysql==1.0.2
python-dotenv==0.19.0
pyjwt==2.1.0
werkzeug==2.0.1